backend/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
pydantic
python-multipart
requests
python-dotenv
networkx
